Command System improvement
