+++ src/main/java/com/massivecraft/factions/cmd/tabcomplete/TabCompleteProvider.java
@@ -10,7 +10,6 @@ import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 public interface TabCompleteProvider {
